Move tile path string to level index width methods from `mysql` to `layout` package
