+++ api/layout/paths.go
@@ -19,6 +19,8 @@ package layout
 
 import (
 	"fmt"
+	"strconv"
+	"strings"
 )
 
 const (
@@ -72,4 +74,65 @@ func fmtN(N uint64) string {
 		N /= 1000
 	}
 	return n
+}
+
+// ParseTileLevelIndexWidth takes level and index in string, validates and returns the level, index and width in uint64.
+//
+// Examples:
+// "/tile/0/x001/x234/067" means level 0 and index 1234067 of a full tile.
+// "/tile/0/x001/x234/067.p/8" means level 0, index 1234067 and width 8 of a partial tile.
+func ParseTileLevelIndexWidth(level, index string) (uint64, uint64, uint64, error) {
+	l, err := ParseTileLevel(level)
+	if err != nil {
+		return 0, 0, 0, err
+	}
+
+	i, w, err := ParseTileIndexWidth(index)
+	if err != nil {
+		return 0, 0, 0, err
+	}
+
+	return l, i, w, err
+}
+
+// ParseTileLevel takes level in string, validates and returns the level in uint64.
+func ParseTileLevel(level string) (uint64, error) {
+	l, err := strconv.ParseUint(level, 10, 64)
+	// Verify that level is an integer between 0 and 63 as specified in the tlog-tiles specification.
+	if l > 63 || err != nil {
+		return 0, fmt.Errorf("failed to parse tile level")
+	}
+	return l, err
+}
+
+// ParseTileIndexWidth takes index in string, validates and returns the index and width in uint64.
+func ParseTileIndexWidth(index string) (uint64, uint64, error) {
+	w := uint64(256)
+	indexPaths := strings.Split(index, "/")
+
+	if strings.Contains(index, ".p") {
+		var err error
+		w, err = strconv.ParseUint(indexPaths[len(indexPaths)-1], 10, 64)
+		if err != nil || w < 1 || w > 255 {
+			return 0, 0, fmt.Errorf("failed to parse tile index")
+		}
+		indexPaths[len(indexPaths)-2] = strings.TrimSuffix(indexPaths[len(indexPaths)-2], ".p")
+		indexPaths = indexPaths[:len(indexPaths)-1]
+	}
+
+	if strings.Count(index, "x") != len(indexPaths)-1 || strings.HasPrefix(indexPaths[len(indexPaths)-1], "x") {
+		return 0, 0, fmt.Errorf("failed to parse tile index")
+	}
+
+	i := uint64(0)
+	for _, indexPath := range indexPaths {
+		indexPath = strings.TrimPrefix(indexPath, "x")
+		n, err := strconv.ParseUint(indexPath, 10, 64)
+		if err != nil || n >= 1000 || len(indexPath) != 3 {
+			return 0, 0, fmt.Errorf("failed to parse tile index")
+		}
+		i = i*1000 + n
+	}
+
+	return i, w, nil
 }
